Enhance cba_config fixture to load an actual config file
Add support for optionally specifying a config file name, located in
the tests/data directory, via a new config marker, defaulting to the
'config_good.yaml` if no marker is specified, which will be loaded
and returned as the result of the fixture.

Added a new session scoped 'data_dir' fixture that returns the path
to the tests/data directory for the current pytest run.