Refactor main() into multiple subroutines
By refectoring the main() routine to call subroutines for various
stages this allows easy unit testing of those stages.

Add unit tests for the new subroutines representing the various
stages of the refactored main.

As part of implementing the test for get_config() the a new fixture,
cba_config_path, was created to provide access to the marked config
path, and the cba_config fixture has been re-worked to leverage it
to determine the path to the file to load.

Also added a unit test for main itself to validate correct operation
of exception handling.

Moved tests/test_adapter.py into tests/unit.

Fixed a typo in tests/unit/test_bill_utils.py