Implement review enhancement suggestions
Move get_billing_dimensions() inside the try/except handling in the
process_metering() call so that if it fails an appropriate error state
is recorded in the csp_config for this iteration of the event loop.

Also ensure that we update the csp_config's timestamp when we are
setting an error.