Incorporate further review feedback
Make the testing for missing dimensions more explicit by testing for
both the case where there is a gap in the dimensions, as well as for
the case where the last dimension entry is not a catch all, meaning
that certain usage levels may not match a specified dimension.

Also refactor the exception handling in process_metering() to remove
some duplication and simpify the code.