Improved initial metering test exception handling
During the initial startup we perform a metering test in an effort
to detect deployment failures as early as possible. If this testing
encounters an error we should endeavour to update the csp_config to
indicate that the error occurred before the adapter process dies.

For the sake of code readability the metering test code has been
moved out of the main() routine to a new metering_test() routine.

Updated existing tests to reflect the move of the metering test to
the new routine, and added an initial test to validate handling of
hook.update_csp_config() failures.

Fixes: #108