Address review feedback
Updated CONTRIBUTING.md to reflect review feedback and to add an
additional note with regard to issues that can be encountered when
using the version of tox (2.9.1) available from openSUSE Leap system
packages.

Retain the explicit running of quality checks and pytest verification
in separate GitHub Actions stages; eliminating the need for using the
tox-gh-actions plugin to control which testenvs run for which Python
version.