Implement supported for tiered consumption reporting
Add support to the get_billing_dimensions() routine for the tiered
consumption reporting type.

Add new exceptions related to invalid consumption reporting type
specification and missing dimensions for the tiered consumption
reporting type.

Update verbiage associated with get_volume_dimensions() to avoid
confusing usage of the word "tier".

Add unit testing support to exercise the new code, including tweaks
to existing testing configs, and the addition of a new config to
exercise the handling of invalid consumption reporting types.

Also add a pragma comment to the adapter.py's call of main() to
indicate that it should be ignored for coverage checking purposes,
allowing the PR to achieve 100% coverage.

Fixes: #120