Add archive bytes limit option to config
This sets the max length of the archive in bytes. If the archive
grows larger than this limit the archive is trimmed until it
satisfies the limit.

For k8s products this limit is 1MiB or 1048576 bytes. If the limit
is set to 0 there is no size limit. The archive can grow as large
as the retention period.